同步修改AC00551-55T-01部分代码

+++ app/src/main/java/com/run/treadmill/ota/BinUpdate.java
@@ -4,10 +4,10 @@ import android.app.Activity;
 import android.app.ProgressDialog;
 
 import com.run.android.ShellCmdUtils;
+import com.run.serial.OTAParam;
 import com.run.serial.SerialUtils;
 import com.run.treadmill.R;
 import com.run.treadmill.manager.ControlManager;
-import com.run.treadmill.manager.SpManager;
 import com.run.treadmill.util.Logger;
 
 import java.io.ByteArrayOutputStream;
@@ -20,11 +20,11 @@ public class BinUpdate extends BaseUpdate {
 
     public void procBin(Activity act, String path) {
         Logger.d("========data======1===");
-        if (SerialUtils.getInstance().isSendBinCnt) {
+        if (OTAParam.isSendBinCnt) {
             return;
         }
         activity = act;
-        SerialUtils.getInstance().isSendBinCnt = true;
+        OTAParam.isSendBinCnt = true;
         SerialUtils.getInstance().reMoveAllReSendPackage();
 
         // 0x6D  0x69
@@ -38,11 +38,11 @@ public class BinUpdate extends BaseUpdate {
             int index = 0;
             boolean isEndSend = false;
 
-            int oneFrameLen = 256;
+            final int oneFrameLen = 256;
 
             @Override
             public void run() {
-                while (!isEndSend && !SerialUtils.isSendBinData) {
+                while (!isEndSend && !OTAParam.isSendBinData) {
                     try {
                         Thread.sleep(1000);
                     } catch (Exception e) {
@@ -56,17 +56,18 @@ public class BinUpdate extends BaseUpdate {
                 }
                 /*SerialUtils.isSendOtaData = true;
                 SerialUtils.isSendOtaOneFrame = true;*/
-                SpManager.setBinUpdate(true);
                 Logger.d("========data======2===");
 
                 while (!isEndSend) {
                     try {
-                        if (SerialUtils.isSendBinData && SerialUtils.isSendBinOneFrame) {
+                        if (OTAParam.reSend) {
+                            OTAParam.isSendBinOneFrame = true;
+                            index = OTAParam.index;
+                        }
+
+                        if (OTAParam.isSendBinData && OTAParam.isSendBinOneFrame) {
                             int readLen = oneFrameLen;
                             byte[] pkgBytes;
-                            /*if ( (index + oneFrameLen) >= data.length ) {
-                                readLen = data.length - index;
-                            }*/
                             pkgBytes = new byte[readLen];
 
                             for (int i = 0; i < pkgBytes.length; i++) {
@@ -80,24 +81,24 @@ public class BinUpdate extends BaseUpdate {
                             }
                             Logger.d("========data====index=====" + index + " data.length " + data.length);
 
-                            SerialUtils.isSendBinOneFrame = false;
+                            OTAParam.isSendBinOneFrame = false;
+
+                            // 保存，用于重发
+                            OTAParam.pkgBytes = pkgBytes;
+                            OTAParam.index = index;
+                            OTAParam.reSend = false;
                             SerialUtils.getInstance().sendOtaDataPackage(pkgBytes, index);
 
                             index += readLen;
-//                            if ( index >= data.length ) {
                             if (index >= 0x0000CF00) {
-                                SpManager.setBinUpdate(false);
                                 Logger.d("========data====3=====");
                                 isEndSend = true;
                                 index = data.length;
                             }
 
-                            activity.runOnUiThread(new Runnable() {
-                                @Override
-                                public void run() {
-                                    //Logger.d("========data====pro=====" + (index * 1.0f / data.length * 100));
-                                    showDialogUpdate((int) (index * 1.0f / 0x0000CF00 * 100));
-                                }
+                            activity.runOnUiThread(() -> {
+                                //Logger.d("========data====pro=====" + (index * 1.0f / data.length * 100));
+                                showDialogUpdate((int) (index * 1.0f / 0x0000CF00 * 100));
                             });
                         }
                         Thread.sleep(400);
@@ -106,9 +107,8 @@ public class BinUpdate extends BaseUpdate {
                     }
                 }
 
-                while (isEndSend && SerialUtils.isSendBinOneFrame) {
+                if (OTAParam.isSendBinOneFrame) {
                     ShellCmdUtils.getInstance().execCommand("reboot");
-                    break;
                 }
             }
         }).start();
